refactor(cli): split interactive.rs into ui/commands/config/data_dir (#1133)
Behavior-preserving module split of the 2,185-line cqlite-cli/src/interactive.rs
(epic #1116). interactive.rs -> interactive/mod.rs + per-concern submodules:

- mod.rs (234): shared model (ReplSession, ColumnInfo, TableInfo) + entry
  points (start_repl_mode, handle_repl_command) + submodule wiring.
- ui.rs (818): prompt/banner/goodbye, :help system, query-result table
  rendering (incl. paging), error hinting, schema/DDL display.
- commands.rs (894): :tables/:keyspaces/:info/:describe/:schema/:use/:history/
  :source dispatch, CQL execution, and database/schema introspection helpers.
- config.rs (134): :config display + mutation.
- data_dir.rs (190): Cassandra data-directory scanning.

No semantic changes. The module remains gated out of the build (main.rs still
has '// mod interactive;'), exactly as before; the split is purely
organizational. Largest module dropped 2,185 -> 894 lines; every file is well
under the ~1,500-line ceiling. Decomposed the >100-line execute_enhanced_cql_query
by extracting its trailing timing/perf/warnings render block into
ui::execute_enhanced_cql_query_render (verbatim, low-risk). Dropped two dead
imports (Cell/Row/Table, ReplConfig/ReplEngine/ReplMode) that were unused in the
original. Validated by temporarily declaring the module: the only compile errors
are the 10 pre-existing stale-cqlite-core-API errors present in the original
file (the original additionally had 2 unused-import errors now removed).
